BackEnd/src/piprequirement.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.9
djangorestframework==3.14.0
django-extensions==3.2.3
idna==3.6
norminette==3.3.55
Pillow==10.2.0
psycopg==3.1.18
psycopg2-binary==2.9.9
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.0
python-dotenv==1.0.1
python-dotenv==1.0.1
pyotp==2.9.0
channels[daphne]==4.0.0
channels-redis==4.1.0